Separate User and bucketlist models
