+++ app/api/tests/test_create_user.py
@@ -10,7 +10,3 @@ class FlaskTestCase(unittest.TestCase):
         tester = app.test_client(self)
         response = tester().get('/homepage', content_type='html/text')
         self.assertEqual(response.status_code, 200)
-
-
-# if __name__ == '__main__':
-#     unittest.main()
